Xiaomi kernel changes for Redmi 3 and Redmi 2L
The Patch based on qcom release TAG LA.BR.1.2.4-04410-8x16.0,
Kernel config for Redmi 3: wt88509_64-perf_defconfig
Kernel config for Redmi 2(Android L): wt88047-perf_defconfig

+++ drivers/video/msm/mdss/mdss_mdp_ctl.c
@@ -959,7 +959,7 @@ int mdss_mdp_perf_bw_check_pipe(struct mdss_mdp_perf_params *perf,
 		return 0;
 
 	vbp_fac = mdss_mdp_get_vbp_factor_max(ctl);
-	prefill_bw = perf->prefill_bytes * vbp_fac;
+	prefill_bw = (u64)(perf->prefill_bytes) * (vbp_fac);
 	pipe_bw = max(prefill_bw, perf->bw_overlap);
 	pr_debug("prefill=%llu, vbp_fac=%u, overlap=%llu\n",
 			prefill_bw, vbp_fac, perf->bw_overlap);
